Upgrade to support new hugo releases.
Hugo from 0.60 upwards is CommonMark compliant.
Apparently that implies that HTML markup is considered unsafe.
Set unsafe=true for now.

After 0.56 symlinks to directories are not supported.
Using module.mounts instead.

With the goldmark renderer the default is to not allow users
to embed HTML snippets into Markdown content, so for now
we have to run with "unsafe = true"

The renderer doesn't like a HTML snippet without a blank
line between that and markdown.

Signed-off-by: Ole Troan <ot@cisco.com>